Code/requirements.txt
# Core Data & Numerical Computing
numpy>=1.24
pandas>=2.0
scipy>=1.10

# Financial Data
yfinance>=0.2.30

# Machine Learning & Statistical Models
scikit-learn>=1.3
hmmlearn>=0.3.0

# Optimization
cvxpy>=1.4

# Visualization
matplotlib>=3.7
seaborn>=0.12

# LLM & Agent Layer
openai>=1.3
python-dotenv>=1.0

# Utilities
pyyaml>=6.0
joblib>=1.3
tqdm>=4.66

#UI
streamlit>=1.32

# Backtesting
backtesting>=0.3.3
Brotli>=1.0.9